Add a benchmark for texts of length 1.
This helps demonstrate the effect of normalization.
text-icu is more efficient because it normalizes Texts
incrementally and can avoid normalizing a whole string
once enough has been read to make a determination of the
ordering.  We can see this by noting that the gap between
our library and text-icu is reduced substantially when
normalization is turned off or when the strings are short
(so that incremental normalization doesn't help).